+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 This changelog is formatted as recommended by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+### Added
+- [Customization] The content of the "Staff" block can be modified
 ### Modified
-- [Customization] The styles are now defined in a config file, no more as attributes in the HTML.
+- [Customization] The styles are now defined in a config file, no more as attributes in the HTML
 
 ## [5.0] - 2025-08-01
 ### Added